Extract sprites
